photo-search-intent-first/requirements.txt
streamlit>=1.32.0
sentence-transformers>=2.7.0
torch>=2.1.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
requests>=2.31.0
openai>=1.52.0
annoy>=1.17.3
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
